beefline-backend/requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9

# Image processing
Pillow==10.1.0

# CORS handling
django-cors-headers==4.3.0

# Environment variables
python-decouple==3.8

# API documentation
drf-spectacular==0.26.5

# Filtering and search
django-filter==23.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Code quality
flake8==6.1.0
black==23.11.0

# Production server
gunicorn==21.2.0
